Added new controllers and services, updated entities and repos

+++ rentalSystem/src/main/java/com/rental/sys/repo/ProductImageRepo.java
@@ -3,6 +3,6 @@ package com.rental.sys.repo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.rental.sys.entities.ProductImage;
 
-public interface ProductImageRepo extends JpaRepository<ProductImage, Long> {
+public interface ProductImageRepo extends JpaRepository<ProductImage, Integer> {
 
 }
